requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-archive==0.2.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
imagekitio==3.0.1
importlib-metadata==6.0.0
Markdown==3.4.1
mysqlclient==2.1.1
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.9.1
sqlparse==0.4.3
urllib3==1.26.15
zipp==3.15.0
